- 스크롤뷰 하이라이트 - 자동 저장 - 삭제시 팝업창

+++ Buildingsurveyapp/app/src/main/java/com/example/building_survey_app/Models/FlawModel.kt
@@ -12,7 +12,7 @@ class FlawModel : Serializable {
     var FlawPos : String =""
     var Flaw : String=""
     var FlawWidth : Double = 0.0
-    var FlawLength : Double = 0.0
+    var FlawLength : Int = 0
     var FlawCount : Int = 0
     var capturedPic : Bitmap? = null
     var compareCapturedPic : Bitmap? = null
